[REFACTOR] magic-login 관련 설정 external.yml로 이동
