+++ src/main/java/sopt/makers/authentication/application/service/auth/VerifyVerificationService.java
@@ -3,9 +3,9 @@ package sopt.makers.authentication.application.service.auth;
 import static sopt.makers.authentication.domain.auth.exception.AuthFailure.EXPIRED_PHONE_VERIFICATION;
 import static sopt.makers.authentication.domain.auth.exception.AuthFailure.INVALID_PHONE_VERIFICATION_CODE;
 
+import sopt.makers.authentication.adapter.out.external.oauth.MagicLoginProperty;
 import sopt.makers.authentication.application.port.in.auth.VerifyPhoneVerificationUsecase;
 import sopt.makers.authentication.application.port.out.auth.PhoneVerificationRepository;
-import sopt.makers.authentication.config.MagicLoginProperty;
 import sopt.makers.authentication.domain.auth.PhoneVerification;
 import sopt.makers.authentication.domain.auth.exception.AuthException;
