+++ src/main/java/sopt/makers/authentication/config/ApplicationConfig.java
@@ -2,6 +2,7 @@ package sopt.makers.authentication.config;
 
 import sopt.makers.authentication.adapter.out.external.oauth.AppleOAuthProperty;
 import sopt.makers.authentication.adapter.out.external.oauth.GoogleOAuthProperty;
+import sopt.makers.authentication.adapter.out.external.oauth.MagicLoginProperty;
 import sopt.makers.authentication.adapter.out.external.sms.GabiaProperty;
 
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
