+++ src/main/resources/application-local.yaml
@@ -7,7 +7,6 @@ spring:
       - classpath:external.yaml
       - classpath:monitoring.yaml
       - classpath:security.yaml
-      - classpath:magic-login.yaml
   datasource:
     url: ${DB_URL}
     username: ${DB_USERNAME}
