+++ src/main/resources/application-prod.yaml
@@ -7,7 +7,6 @@ spring:
       - classpath:external.yaml
       - classpath:security.yaml
       - classpath:monitoring.yaml
-      - classpath:magic-login.yaml
   datasource:
     url: ${DB_URL}
     username: ${DB_USERNAME}
