+++ src/main/resources/application-test.yaml
@@ -6,7 +6,6 @@ spring:
       - classpath:jpa.yaml
       - classpath:external.yaml
       - classpath:security.yaml
-      - classpath:magic-login.yaml
   datasource:
     url: ${DB_URL}
     username: ${DB_USERNAME}
