+++ src/main/resources/external.yaml
@@ -17,3 +17,7 @@ external:
     google:
       client:
         id: ${GOOGLE_CLIENT_ID}
+    magic-login:
+      phone: ${MAGIC_LOGIN_PHONE}
+      code: ${MAGIC_LOGIN_CODE}
+      name: ${MAGIC_LOGIN_NAME}
